Add a quick hack for Youtube stills
YouTube has an issue with the :empty requirement on tags that include
background images via CSS. It breaks the uninversion of the "videowall"
stills at the end of videos. This is a quick and dirty hack to fix it
until I can come up with a better solution in general to the
double-inversion problem.