Uninvert gmail thread avatars in smart mode
Smart mode was formerly using the presence of `aria-hidden="true"`
attribute as a signal to leave the image alone, as used for math markup
on Wikipedia. This unfortunately included the gmail avatars, but those
don't have alt text, so I'm now using alt text in the match rule. Also
narrowing the scope of the "aria-hidden with alt text" rule to just img
tags.